+++ core/src/com/unciv/logic/automation/NextTurnAutomation.kt
@@ -115,11 +115,11 @@ object NextTurnAutomation {
     private fun tryGainInfluence(civInfo: CivilizationInfo, cityState: CivilizationInfo) {
         if (civInfo.gold < 250) return // save up
         if (cityState.getDiplomacyManager(civInfo).influence < 20) {
-            civInfo.giveGoldGift(cityState, 250)
+            cityState.receiveGoldGift(civInfo, 250)
             return
         }
         if (civInfo.gold < 500) return // it's not worth it to invest now, wait until you have enough for 2
-        civInfo.giveGoldGift(cityState, 500)
+        cityState.receiveGoldGift(civInfo, 500)
         return
     }
 
@@ -169,7 +169,7 @@ object NextTurnAutomation {
 
     private fun valueCityStateAlliance(civInfo: CivilizationInfo, cityState: CivilizationInfo): Int {
         var value = 0
-        if (!cityState.isAlive() || cityState.cities.isEmpty())
+        if (!cityState.isAlive() || cityState.cities.isEmpty() || civInfo.cities.isEmpty())
             return value
 
         if (civInfo.victoryType() == VictoryType.Cultural && cityState.canGiveStat(Stat.Culture)) {
@@ -230,9 +230,9 @@ object NextTurnAutomation {
                     && valueCityStateAlliance(civInfo, state) <= 0
                     && state.getTributeWillingness(civInfo) > 0) {
                 if (state.getTributeWillingness(civInfo, demandingWorker = true) > 0)
-                    civInfo.demandWorker(state)
+                    state.tributeWorker(civInfo)
                 else
-                    civInfo.demandGold(state)
+                    state.tributeGold(civInfo)
             }
         }
     }
@@ -328,16 +328,12 @@ object NextTurnAutomation {
     }
     
     private fun foundReligion(civInfo: CivilizationInfo) {
-        println("Founding check?")
         if (civInfo.religionManager.religionState != ReligionState.FoundingReligion) return
-        println("Founding check!!")
         val religionIcon = civInfo.gameInfo.ruleSet.religions
-            .filterNot { civInfo.gameInfo.religions.values.map { religion -> religion.iconName }.contains(it) }
+            .filterNot { civInfo.gameInfo.religions.values.map { religion -> religion.name }.contains(it) }
             .randomOrNull()
             ?: return // Wait what? How did we pass the checking when using a great prophet but not this?
-        println("Icon has been chosen")
         val chosenBeliefs = chooseBeliefs(civInfo, civInfo.religionManager.getBeliefsToChooseAtFounding()).toList()
-        println("Beliefs have been chosen")
         civInfo.religionManager.chooseBeliefs(religionIcon, religionIcon, chosenBeliefs)
     }
     
@@ -488,7 +484,7 @@ object NextTurnAutomation {
 
             val unitsInBorder = otherCiv.getCivUnits().count { !it.isCivilian() && it.getTile().getOwner() == civInfo }
             if (unitsInBorder > 0 && diplomacy.relationshipLevel() < RelationshipLevel.Friend) {
-                diplomacy.influence -= 10f
+                diplomacy.addInfluence(-10f)
                 if (!diplomacy.hasFlag(DiplomacyFlags.BorderConflict)) {
                     otherCiv.popupAlerts.add(PopupAlert(AlertType.BorderConflict, civInfo.civName))
                     diplomacy.setFlag(DiplomacyFlags.BorderConflict, 10)