+++ core/src/com/unciv/logic/map/MapUnit.kt
@@ -101,15 +101,17 @@ class MapUnit {
 
     /**
      * Name which should be displayed in UI
-     * 
+     *
      * Note this is translated after being returned from this function, so let's pay
      * attention to combined names (renamed units, religion).
      */
     fun displayName(): String {
-        val name = if (instanceName == null) name
-                   else "$instanceName ({${name}})"
-        return if (religion != null) "[$name] ([$religion])"
-               else name
+        val baseName =
+            if (instanceName == null) "[$name]"
+            else "instanceName ([$name])"
+
+        return if (religion == null) baseName
+        else "$baseName ([${getReligionDisplayName()}])"
     }
 
     var currentMovement: Float = 0f
@@ -168,9 +170,10 @@ class MapUnit {
 
     fun getTile(): TileInfo = currentTile
     fun getMaxMovement(): Int {
-        if (isEmbarked()) return getEmbarkedMovement()
-
-        var movement = baseUnit.movement
+        var movement = 
+            if (isEmbarked()) 2
+            else baseUnit.movement
+        
         movement += getMatchingUniques("[] Movement").sumBy { it.params[0].toInt() }
 
         for (unique in civInfo.getMatchingUniques("+[] Movement for all [] units"))
@@ -181,6 +184,13 @@ class MapUnit {
             civInfo.hasUnique("+1 Movement for all units during Golden Age")
         )
             movement += 1
+
+        // Deprecated since 3.16.11
+            if (isEmbarked()) {
+                movement += civInfo.getMatchingUniques("Increases embarked movement +1").count()
+                if (civInfo.hasUnique("+1 Movement for all embarked units")) movement += 1
+            }
+        //
 
         return movement
     }
@@ -194,10 +204,11 @@ class MapUnit {
     fun getUniques(): ArrayList<Unique> = tempUniques
 
     fun getMatchingUniques(placeholderText: String): Sequence<Unique> =
-        tempUniques.asSequence().filter { it.placeholderText == placeholderText }
+        tempUniques.asSequence().filter { it.placeholderText == placeholderText } + 
+            civInfo.getMatchingUniques(placeholderText)
 
     fun hasUnique(unique: String): Boolean {
-        return getUniques().any { it.placeholderText == unique }
+        return getUniques().any { it.placeholderText == unique } || civInfo.hasUnique(unique)
     }
 
     fun updateUniques() {
@@ -341,7 +352,6 @@ class MapUnit {
         return range
     }
 
-
     fun isEmbarked(): Boolean {
         if (!baseUnit.isLandUnit()) return false
         return currentTile.isWater
@@ -355,13 +365,6 @@ class MapUnit {
                 it.getOwner() == to || it.getUnits().any { unit -> unit.owner == to.civName }
             }
         return false
-    }
-
-    fun getEmbarkedMovement(): Int {
-        var movement = 2
-        movement += civInfo.getMatchingUniques("Increases embarked movement +1").count()
-        if (civInfo.hasUnique("+1 Movement for all embarked units")) movement += 1
-        return movement
     }
 
     fun getUnitToUpgradeTo(): BaseUnit {
@@ -966,6 +969,11 @@ class MapUnit {
     fun canBuildImprovement(improvement: TileImprovement, tile: TileInfo = currentTile): Boolean {
         val matchingUniques = getMatchingUniques(Constants.canBuildImprovements)
         return matchingUniques.any { improvement.matchesFilter(it.params[0]) || tile.matchesTerrainFilter(it.params[0]) }
+    }
+    
+    fun getReligionDisplayName(): String? {
+        if (religion == null) return null
+        return civInfo.gameInfo.religions[religion]!!.displayName ?: religion
     }
     
     fun religiousActionsUnitCanDo(): Sequence<String> {